Fixed Todo's
